+++ src/Contract.Common/Dtos/OrderDtos/OrderDto.cs
@@ -95,7 +95,7 @@ public class OrderDto
     public string? VoidByName { get; set; }
     public DateTime? VoidTime { get; set; }
     public string? VoidReason { get; set; }
-    public int? VoidCount { get; set; }
+    public decimal? VoidCount { get; set; }
     public decimal? TotalVoid { get; set; }
 
     //Print Details
